requirements.txt
confluent-kafka>=2.8.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
requests>=2.32.0
websockets>=14.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
apscheduler>=3.10.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
pandas>=2.0.0
pytest>=8.0.0
httpx>=0.27.0
